Consolidate badge stories into badges.stories.jsx
